edit updated playlist in storage

+++ src/components/PlayListManager/index.tsx
@@ -12,18 +12,21 @@ import "./style.scss";
 const clearModel: listOfPlaylistItemType = {
 	name: "",
 	url: "",
+	description: "",
 };
 
 export type propTypes = {
 	onUpdatePlaylist(url: string | undefined): void;
 	onSetCurrentPlaylistNumber(number: number): void;
 	onSetList(playList: listOfPlaylistItemType[]): void;
+	onGetPleyListFromStorage(): void;
 };
 
 const PlayListManager = ({
 	onSetList,
-	onSetCurrentPlaylistNumber,
 	onUpdatePlaylist,
+	onGetPleyListFromStorage,
+	onSetCurrentPlaylistNumber,
 }: propTypes) => {
 	// const [listOfPlaylist, props.onSetList] = React.useState<listOfPlaylistItemType[]>([]);
 	const [selectedItem, setSelectedItem] = React.useState<listOfPlaylistItemType>(clearModel);
@@ -33,30 +36,41 @@ const PlayListManager = ({
 
 	const mainContext: mainContextType = React.useContext<mainContextType>(MainContext);
 
-	const handleAddItem = (newItem: listOfPlaylistItemType) =>
+	const handleAddItem = (newItem: listOfPlaylistItemType) => {
+		handleSavePlaylistToStorage(newItem);
 		onSetList([...mainContext.listOfPlaylist, newItem]);
+		onGetPleyListFromStorage();
+	};
 
 	const handleChangeItem = ({ target }: { target: { id: string; value: string } }) =>
 		setSelectedItem({ ...selectedItem, [target.id]: target.value });
 
 	const handleRemoveItem = (removingIndex: number) => {
-		const removingItem = mainContext.listOfPlaylist.find(
-			(_: any, index: number) => index === removingIndex,
-		);
+		if (removingIndex >= 0) {
+			const removingItem = mainContext.listOfPlaylist.find(
+				(_: any, index: number) => index === removingIndex,
+			);
 
-		db.removeData("playLists", { name: removingItem.name, url: removingItem.url });
+			db.removeData("playLists", { name: removingItem.name, url: removingItem.url });
 
-		onSetList([
-			...mainContext.listOfPlaylist.filter((_: any, index: number) => index !== removingIndex),
-		]);
+			onSetList([
+				...mainContext.listOfPlaylist.filter((_: any, index: number) => index !== removingIndex),
+			]);
+		}
 	};
 
-	const handleUpdateItem = (newItem: listOfPlaylistItemType, indexOfItem: number) =>
-		onSetList([
-			...mainContext.listOfPlaylist.map((item: listOfPlaylistItemType, index: number) =>
-				index === indexOfItem ? newItem : item,
-			),
-		]);
+	const handleUpdateItem = async (newItem: listOfPlaylistItemType, indexOfItem: number) => {
+		if (newItem.id) {
+			await db.updateData("playLists", { id: newItem.id }, newItem);
+			onGetPleyListFromStorage();
+		} else {
+			onSetList([
+				...mainContext.listOfPlaylist.map((item: listOfPlaylistItemType, index: number) =>
+					index === indexOfItem ? newItem : item,
+				),
+			]);
+		}
+	};
 
 	const handleSavePlaylistToStorage = (playlist: listOfPlaylistItemType): void => {
 		db.setData("playLists", playlist);
@@ -64,17 +78,18 @@ const PlayListManager = ({
 
 	const handleSave = () => {
 		if (indexForListEditor === null) {
-			handleSavePlaylistToStorage(selectedItem);
 			handleAddItem(selectedItem);
 		} else if (indexForListEditor) {
 			handleUpdateItem(selectedItem, indexForListEditor);
 		}
+
 		// closing form for input
-		setIndexForListEditor(indexForListEditor === undefined ? null : undefined);
 		setSelectedItem(clearModel);
+		setIndexForListEditor(indexForListEditor === undefined ? null : undefined);
 	};
 
 	const handleAddPlaylist = () => {
+		setSelectedItem(clearModel);
 		setIndexForListEditor(indexForListEditor === undefined ? null : undefined);
 	};
 
@@ -84,7 +99,17 @@ const PlayListManager = ({
 			onUpdatePlaylist(mainContext.listOfPlaylist[index].url);
 	};
 
-	const handleSelectItem = (index: number) => console.log(mainContext.listOfPlaylist[index]); // onSelectItem(listOfPlaylist[index]);
+	const handleSelectItem = (index: number) => {
+		// console.log(mainContext.listOfPlaylist[index]); // onSelectItem(listOfPlaylist[index]);
+	};
+	const handleEditItem = (index: number) => {
+		if (Array.isArray(mainContext.listOfPlaylist) && mainContext.listOfPlaylist.length) {
+			const selectedItem = mainContext.listOfPlaylist[index];
+			handleAddPlaylist();
+			setIndexForListEditor(index);
+			setSelectedItem(selectedItem);
+		}
+	};
 
 	return (
 		<div id="component-listOfPlaylistItemType">
@@ -102,11 +127,18 @@ const PlayListManager = ({
 							<span>{playlistItem.description}</span>
 						</a>
 						<button
+							className="component-listOfPlaylistItemType__edit-button"
+							type="button"
+							onClick={() => handleEditItem(index)}
+						>
+							<i className="far fa-edit"></i>
+						</button>{" "}
+						<button
 							className="component-listOfPlaylistItemType__remove-button"
 							type="button"
 							onClick={() => handleRemoveItem(index)}
 						>
-							-
+							<i className="far fa-trash-alt"></i>
 						</button>
 					</li>
 				))}
@@ -123,16 +155,29 @@ const PlayListManager = ({
 			{indexForListEditor !== undefined && (
 				<div className="component-listOfPlaylistItemType-editor">
 					<div>
-						<input type="text" id="url" placeholder="Url" onChange={handleChangeItem} />
-						<input type="text" id="name" placeholder="Name" onChange={handleChangeItem} />
+						<input
+							type="text"
+							id="url"
+							placeholder="Url"
+							onChange={handleChangeItem}
+							value={selectedItem.url}
+						/>
+						<input
+							type="text"
+							id="name"
+							placeholder="Name"
+							onChange={handleChangeItem}
+							value={selectedItem.name}
+						/>
 						<input
 							type="text"
 							id="description"
 							placeholder="Description"
 							onChange={handleChangeItem}
+							value={selectedItem.description}
 						/>
 					</div>
-					<button type="button" onClick={handleSave}>
+					<button type="button" onClick={() => handleSave()}>
 						Save
 					</button>
 				</div>
